Pick dependencies from SwiftPM
- PartoutRuntime from partout (header only)
- PartoutNative from app-apple (XCFramework)
