+++ app-apple/Package.swift
@@ -31,6 +31,10 @@ let package = Package(
             targets: ["AppLibraryTV"]
         ),
         .library(
+            name: "PartoutNative",
+            targets: ["PartoutNative"]
+        ),
+        .library(
             name: "TunnelLibrary",
             targets: ["TunnelLibrary"]
         )
@@ -81,6 +85,10 @@ let package = Package(
             resources: [
                 .process("Resources")
             ]
+        ),
+        .binaryTarget(
+            name: "PartoutNative",
+            path: "PartoutNative.xcframework"
         ),
         .target(
             name: "TunnelLibrary",
