--- app-apple/Sources/Empty/AppPartoutRuntime/Dummy.swift
@@ -1,3 +1,0 @@
-// SPDX-FileCopyrightText: 2026 Davide De Rosa
-//
-// SPDX-License-Identifier: GPL-3.0
