+++ app-apple/Sources/TunnelLibrary/TunnelBackendProtocol.swift
@@ -4,7 +4,6 @@
 
 import CommonLibrary
 import Partout
-import PartoutRuntime
 
 public protocol TunnelBackendProtocol: AnyObject, Sendable {
     func start() async throws
@@ -21,23 +20,5 @@ extension SimpleConnectionDaemon: TunnelBackendProtocol {
         let input = try ABI.decode(Message.Input.self, from: messageData)
         let output = try await sendMessage(input)
         return try ABI.encode(output)
-    }
-}
-
-extension PartoutProviderRuntime: TunnelBackendProtocol {
-    public func start() async throws {
-        try await startTunnel()
-    }
-
-    public func stop() async {
-        await stopTunnel()
-    }
-
-    public func hold() async {
-        await holdTunnel()
-    }
-
-    public func sendMessage(_ messageData: Data) async throws -> Data? {
-        await handleAppMessage(messageData)
     }
 }
